認証・認可

+++ chapter-8/src/app/admin/categories/[id]/page.tsx
@@ -4,6 +4,8 @@ import { useCategories } from "@/app/_hooks/useCategories";
 import { useParams } from "next/navigation";
 import CategoryForm from "@/app/_components/CategoryForm";
 import { useRouter } from "next/navigation";
+import { useSupabaseSession } from "@/app/_hooks/useSupabaseSession";
+
 export default function Home() {
   const title = "カテゴリー編集";
   const Params = useParams();
@@ -11,9 +13,15 @@ export default function Home() {
   const categoryId = Number(Params.id);
   const { categories } = useCategories();
   const category = categories.find((category) => category.id === categoryId);
+  const { token } = useSupabaseSession();
+
   const handleSubmit = async (categoryName: string) => {
+    if (!token) return;
     const res = await fetch(`/api/admin/categories/${categoryId}`, {
       method: "PUT",
+      headers: {
+        Authorization: token,
+      },
       body: JSON.stringify({ name: categoryName }),
     });
     if (res.ok) {
@@ -22,9 +30,13 @@ export default function Home() {
     }
   };
   const handleDelete = async () => {
+    if (!token) return;
     if (!confirm("削除しますか？")) return;
     await fetch(`/api/admin/categories/${categoryId}`, {
       method: "DELETE",
+      headers: {
+        Authorization: token,
+      },
     });
     router.push("/admin/categories");
     router.refresh();
